feat : process 부분 photo 분리
