+++ src/main/java/com/example/healthgenie/boundedContext/process/photo/dto/ProcessPhotoDeleteResponse.java
@@ -1,0 +1,11 @@
+package com.example.healthgenie.boundedContext.process.photo.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+
+@Getter
+@Builder
+public class ProcessPhotoDeleteResponse {
+    private Long id;
+}
