+++ src/main/java/com/example/healthgenie/boundedContext/process/process/dto/PtProcessDeleteResponseDto.java
@@ -1,0 +1,10 @@
+package com.example.healthgenie.boundedContext.process.process.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class PtProcessDeleteResponseDto {
+    private Long id;
+}
