+++ src/main/java/com/example/healthgenie/boundedContext/process/process/repository/PtProcessRepository.java
@@ -1,0 +1,8 @@
+package com.example.healthgenie.boundedContext.process.process.repository;
+
+import com.example.healthgenie.boundedContext.process.process.entity.PtProcess;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PtProcessRepository extends JpaRepository<PtProcess, Long> {
+
+}
