--- src/main/java/com/example/healthgenie/boundedContext/ptrecord/dto/PtProcessDeleteResponseDto.java
@@ -1,10 +1,0 @@
-package com.example.healthgenie.boundedContext.ptrecord.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class PtProcessDeleteResponseDto {
-    private Long id;
-}
