--- src/main/java/com/example/healthgenie/boundedContext/ptrecord/repository/PtProcessPhotoRepository.java
@@ -1,9 +1,0 @@
-package com.example.healthgenie.boundedContext.ptrecord.repository;
-
-import com.example.healthgenie.boundedContext.ptrecord.entity.PtProcessPhoto;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-public interface PtProcessPhotoRepository extends JpaRepository<PtProcessPhoto, Long> {
-
-}
