--- src/main/java/com/example/healthgenie/boundedContext/ptreview/dto/PtReviewDeleteResponseDto.java
@@ -1,10 +1,0 @@
-package com.example.healthgenie.boundedContext.ptreview.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Getter
-@Builder
-public class PtReviewDeleteResponseDto {
-    private Long id;
-}
