--- src/main/java/com/example/healthgenie/boundedContext/ptreview/repository/PtReviewRepository.java
@@ -1,10 +1,0 @@
-package com.example.healthgenie.boundedContext.ptreview.repository;
-
-import com.example.healthgenie.boundedContext.ptreview.entity.PtReview;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface PtReviewRepository extends JpaRepository<PtReview, Long> {
-
-
-    PtReview findByMemberIdAndTrainerId(Long userId, Long trainerId);
-}
