+++ src/main/java/com/example/healthgenie/boundedContext/review/dto/PtReviewDeleteResponseDto.java
@@ -1,0 +1,10 @@
+package com.example.healthgenie.boundedContext.review.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Getter
+@Builder
+public class PtReviewDeleteResponseDto {
+    private Long id;
+}
