+++ src/main/java/com/example/healthgenie/boundedContext/trainer/photo/dto/ProfilePhotoRequest.java
@@ -1,5 +1,7 @@
 package com.example.healthgenie.boundedContext.trainer.photo.dto;
 
+import com.example.healthgenie.boundedContext.trainer.photo.entity.TrainerPhoto;
+import com.example.healthgenie.boundedContext.trainer.profile.entity.TrainerInfo;
 import jakarta.validation.constraints.Size;
 import java.util.List;
 import lombok.AllArgsConstructor;
@@ -13,6 +15,14 @@ import org.springframework.web.multipart.MultipartFile;
 @NoArgsConstructor
 @Builder
 public class ProfilePhotoRequest {
-    @Size(max = 3)
+    @Size(max = 5)
     private List<MultipartFile> photos;
+
+    public TrainerPhoto toEntity(TrainerInfo info, String uploadUrl, String originName) {
+        return TrainerPhoto.builder()
+                .info(info)
+                .infoPhotoPath(uploadUrl)
+                .name(originName)
+                .build();
+    }
 }
