+++ src/main/java/com/example/healthgenie/boundedContext/trainer/photo/service/ProfilePhotoService.java
@@ -22,7 +22,6 @@ import org.springframework.web.multipart.MultipartFile;
 @Service
 @Slf4j
 @RequiredArgsConstructor
-@Transactional(readOnly = true)
 public class ProfilePhotoService {
 
     private final ProfileRepository profileRepository;
@@ -37,59 +36,19 @@ public class ProfilePhotoService {
         TrainerInfo profile = profileRepository.findById(profileId)
                 .orElseThrow(() -> CustomException.TRAINER_INFO_EMPTY);
 
-        check_permission(userId, profile);
+        checkPermission(userId, profile);
 
         for (MultipartFile file : dto.getPhotos()) {
             String uploadUrl = s3UploadUtils.upload(file, "trainer-profile-photos");
             String originName = file.getOriginalFilename();
 
             TrainerPhoto savedPhoto = trainerProfilePhotoRepository.save(
-                    TrainerPhoto.builder()
-                            .infoPhotoPath(uploadUrl)
-                            .info(profile)
-                            .name(originName)
-                            .build()
-            );
+                    dto.toEntity(profile, uploadUrl, originName));
 
             photos.add(savedPhoto);
         }
 
         return photos;
-    }
-
-    @Transactional
-    public List<TrainerPhoto> saveAll(Long profileId, List<String> photoPaths) {
-        TrainerInfo profile = profileRepository.findById(profileId)
-                .orElseThrow(() -> CustomException.TRAINER_INFO_EMPTY);
-
-        List<TrainerPhoto> photos = photoPaths.stream()
-                .map(path -> TrainerPhoto.builder()
-                        .infoPhotoPath(path)
-                        .info(profile)
-                        .build())
-                .toList();
-
-        // 객체 그래프 탐색용
-        for (TrainerPhoto photo : photos) {
-            profile.addPhoto(photo);
-        }
-
-        return trainerProfilePhotoRepository.saveAll(photos);
-    }
-
-    @Transactional
-    public List<TrainerPhoto> updateAll(Long profileId, List<String> photoPaths) {
-        TrainerInfo info = profileRepository.findById(profileId)
-                .orElseThrow(() -> CustomException.TRAINER_INFO_EMPTY);
-
-        // 객체 그래프 탐색용
-        info.removePhotos(info.getTrainerPhotos());
-
-        // 기존 Photo 삭제
-        trainerProfilePhotoRepository.deleteAllByInfoId(profileId);
-
-        // 새로운 Photo 저장
-        return saveAll(profileId, photoPaths);
     }
 
     @Transactional(readOnly = true)
@@ -108,7 +67,7 @@ public class ProfilePhotoService {
         TrainerInfo profile = profileRepository.findById(profileId)
                 .orElseThrow(() -> CustomException.TRAINER_INFO_EMPTY);
 
-        check_permission(userId, profile);
+        checkPermission(userId, profile);
 
         List<TrainerPhoto> photos = findAllByProfileId(profileId);
 
@@ -125,7 +84,7 @@ public class ProfilePhotoService {
                 .build();
     }
 
-    private static void check_permission(Long userId, TrainerInfo profile) {
+    private static void checkPermission(Long userId, TrainerInfo profile) {
         if (!Objects.equals(userId, profile.getMember().getId())) {
             throw CustomException.NO_PERMISSION;
         }
